[*] Adicionado todos as versões da ABRSAF v2. [*] Ajustado as classes de acordo com a versão correta da ABRSAF v2.
