+++ src/ACBr.Net.NFSe.DANFSe.FastReport/ACBr.Net.NFSe.DANFSe.FastReport.csproj
@@ -79,6 +79,9 @@
     <PackageReference Include="ACBr.Net.DFe.Core">
       <Version>1.2.8.2</Version>
     </PackageReference>
+    <PackageReference Include="FastReport.Net.Demo">
+      <Version>2021.2.7</Version>
+    </PackageReference>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Properties\Resources.resx">
